Use back matter for diagram
Based on the specification's [wording][1] it seems like it is
permissible to directly link to documents (when `rel` isn't `reference`)
but the rules enforced by `metaschema` and `liboscaljava` seem to differ
and require it to start with a `#` (be in internal reference). That's
fine because we can just move it to back matter.

I will review the spec further and perhaps open an issue upstream if
this is a discrepancy in the implementation.