Update OSCAL content based on the changes on NIST's content
Unsure if the updates that I made were the based of the correct
version of the OSCAL content on NIST's GitHub so I linked the sources
of where I made the changes according to.

example Component definition (MongoDB)
https://github.com/usnistgov/oscal-content/blob/main/examples/component-definition/json/example-component-definition.json

profile 4
https://github.com/usnistgov/oscal-content/commit/159073e8a8ca16b3f8fba695aa00018f12535719#diff-7da3dc24b2ec3bfb34e6b00357d91153549394cc42fc8097b963bf71db926d5a

profile 5
https://github.com/usnistgov/oscal-content/blob/main/nist.gov/SP800-53/rev5/json/NIST_SP-800-53_rev5_MODERATE-baseline_profile.json

These links are the where I got the changes from.
In short these are the changes which I made
Component (MongoDB)
Line 70

Profile 4
Line 2569

Profile 5
Line 59