Make formatting match NIST's demo content
Many of the differences found in the *diff.txt files were there
due to differences in spacing or formatting of code.
This commit addresses those differences.

Note: SSP has some changes that involve changes from the NIST ssp
content. It is still work in progress. Including all of the changes
in ssp + format differences would have be very hard to review.

This commit will make comparing the differences easier.