chore: add schema validation in VS Code
This adds support for schema validation in VS Code (and other tools that
use its JSON language server). The underlying schema does not allow the
`$schema` field in objects so editor configuration is the best that we
can do. This makes sure to use the same schemas used in the pipeline.

For now, we follow the `main` branch of upstream but if more significant
changes start to occur between releases, we can switch to a tagged
version at any time.